+++ src/Pages/Login/LoginTap/LoginTap.scss
@@ -6,13 +6,13 @@
 
   .category {
     all: revert;
-    background-color: #76ba99;
-    border-top-left-radius: 15px;
-    border-top-right-radius: 15px;
     height: 50px;
     width: 194px;
-    border: #76ba99 solid;
     color: white;
+    background-color: #76ba99;
+    border: #76ba99 solid;
+    border-top-left-radius: 15px;
+    border-top-right-radius: 15px;
     font-weight: bolder;
   }
 }
